move validator interfaces to their own file
